+++ src/redux/reducers/accountsReducer.js
@@ -1,6 +1,7 @@
 import { handleActions } from 'redux-actions';
 import {
   GET_ACCOUNT_INFO,
+  GET_ACCOUNT_ASSETS,
   GET_ACCOUNT_TXS,
   GET_ACCOUNT_DELEGATIONS,
   GET_ACCOUNT_REDELEGATIONS,
@@ -17,7 +18,7 @@ export const initialState = {
   accountTxsLoading: false,
   accountTxs: [],
   accountTxsPages: 0,
-  // Account Assets (Get brought in with account info)
+  // Account Assets
   accountAssetsLoading: false,
   accountAssets: [],
   accountAssetsPages: 0,
@@ -48,22 +49,42 @@ const reducer = handleActions(
       return {
         ...state,
         accountInfoLoading: true,
-        accountAssetsLoading: true,
       };
     },
     [`${GET_ACCOUNT_INFO}_${SUCCESS}`](state, { payload: accountInfo }) {
       return {
         ...state,
         accountInfoLoading: false,
-        accountAssetsLoading: false,
         accountInfo,
-        accountAssets: accountInfo.balances || [],
       };
     },
     [`${GET_ACCOUNT_INFO}_${FAILURE}`](state) {
       return {
         ...state,
         accountInfoLoading: false,
+      };
+    },
+    /* -----------------
+    GET_ACCOUNT_ASSETS
+    -------------------*/
+    [`${GET_ACCOUNT_ASSETS}_${REQUEST}`](state) {
+      return {
+        ...state,
+        accountAssetsLoading: true,
+      };
+    },
+    [`${GET_ACCOUNT_ASSETS}_${SUCCESS}`](state, { payload }) {
+      const { pages: accountAssetsPages, results: accountAssets = [] } = payload;
+      return {
+        ...state,
+        accountAssetsLoading: false,
+        accountAssets,
+        accountAssetsPages,
+      };
+    },
+    [`${GET_ACCOUNT_ASSETS}_${FAILURE}`](state) {
+      return {
+        ...state,
         accountAssetsLoading: false,
       };
     },