+++ src/utils/number/nHashtoHash.js
@@ -1,3 +1,4 @@
+import { currencyFormat } from './currencyFormat';
 import { numberFormat } from './numberFormat';
 
 export const formatNhash = (nhash, options = {}) => {
@@ -8,7 +9,8 @@ export const formatNhash = (nhash, options = {}) => {
   // If it's not a number or string kick it back
   if (!isNumber && !isString) return nhash;
   // If nhash is given as a string, convert it to a number
-  const nhashFinal = isString ? Number(nhash) : nhash;
+  // const nhashFinal = isString ? Number(nhash) : nhash;
+  const nhashFinal = currencyFormat(nhash, 'nhash');
   const { shorthand = false, decimal = 7 } = options;
   // Convert to formatted number with decimal places and optional shorthand
   return numberFormat(nhashFinal, decimal, { shorthand });